pyproject.toml
[project]
name = "tap-googleadsservice"
version = "0.0.1"
description = "Singer tap for googleadsservice, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Evan Marrow", email = "evan.marrow1@gmail.com" }]
keywords = [
    "ELT",
    "googleadsservice",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license-files = [ "LICENSE" ]
requires-python = ">=3.9"
dependencies = [
    "google-ads>=26.1.0",
    "singer-sdk~=0.46.0",
]

[project.optional-dependencies]
s3 = [
    "fs-s3fs~=1.1.1",
]

[project.scripts]
# CLI declaration
tap-googleadsservice = 'tap_googleadsservice.tap:Tapgoogleadsservice.cli'

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "pytest>=8",
    "pytest-github-actions-annotate-failures>=0.3",
    "singer-sdk[testing]",
]

[tool.pytest.ini_options]
addopts = [
    "--durations=10",
]

[tool.mypy]
warn_unused_configs = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
ignore = [
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = [
    "hatchling>=1,<2",
]
build-backend = "hatchling.build"
